# Run toml-test with "ruby --disble-gems"