@mame 2025-04-22
Thank you. However, regular gems should not use `--disable-gems`. It's clearly stated in `ruby --help` that `--disable-gems` is "only for debugging":

```
Features:
  gems            Rubygems (only for debugging, default: enabled).
```

If startup time for decoder.rb and encoder.rb is a problem, I think it would be better to consider turning them into a server.