@arp242 2025-04-22
Right, that's not mentioned in the manpage 😅

I'm somewhat frequently running this as part of https://toml-lang.github.io/toml-test-matrix, and since that's running toml-test on a few dozen implementations, shaving off a few seconds here and there adds up.

I can also add the flag via a wrapper script, doesn't strictly need to be in the hashbang, although personally I think it should be okay for an internal tool like this (which isn't really a "gem") and having tests run faster for everyone seems nice. But it's your choice so feel free to close if you don't want to.